// 03_Resources/09_Platzi/00_Side_Project/Oil/drilling-calculator/src/store/drilling-types.ts
export interface VolumetricsResult {
  holeCapacity: number;
  drillPipeCapacity: number;
  hwdpCapacity: number;
  dcCapacity: number;
  annularDP: number;
  annularHWDP: number;
  annularDC: number;
  displacementDP: number;
  displacementHWDP: number;
  displacementDC: number;
  volumeInsideDP: number;
  volumeInsideHWDP: number;
  volumeInsideDC: number;
  totalInsideVolume: number;
  volumeAnnularDP: number;
  volumeAnnularHWDP: number;
  volumeAnnularDC: number;
  totalAnnularVolume: number;
  totalSystemVolume: number;
  openHoleVolume: number;
  surfaceToBitTime: number;
  bottomsUpTime: number;
  totalCirculationTime: number;
}

export interface RheologyResult {
  pv: number;
  yp: number;
  av: number;
  pvYpRatio: number;
  n_pl: number;
  k_pl: number;
  n_hb: number;
  k_hb: number;
  tau0_hb: number;
  mu_eff: number;
  gel10s: number;
  gel10m: number;
  gelProgression: number;
}

export interface PumpResult {
  outputPerStroke: number;
  flowRateGPM: number;
  flowRateBBLmin: number;
  hydraulicHP: number;
}